+++ R/checkers.R
@@ -68,11 +68,9 @@ are_calls <- function(...) {
   all(map_vec(enexprs(...), is.call))
 }
 
-search_ <- function(x, callback, na_rm = TRUE, n = 1) {
-  if (na_rm) {
-    x <- vec_drop_na(x)
-  }
-  have_passed <- !callback(x)
+search_ <- function(x, callback, n = 1) {
+  x <- vec_drop_na(x)
+  have_passed <- if (missing(callback)) !x else !callback(x)
   if (all(have_passed)) {
     return()
   }
@@ -243,11 +241,12 @@ check_args <- function(type, x, ...) {
   })
 }
 
-check_suffix_format <- function(x, allowed, arg = caller_arg(x)) {
+check_suffix_format <- function(x, arg = caller_arg(x)) {
   check_string(x, allow_null = TRUE, arg = arg)
   if (is.null(x)) {
     return(invisible(NULL))
   }
+  allowed <- c("a", "c", "n", "o", "^", ",")
   pattern <- to_chr_class(allowed, negate = TRUE)
   keys <- als_extract_keys(x)
   has_dup_keys <- vec_duplicate_any(keys)