+++ R/status-setter.R
@@ -58,8 +58,8 @@ StatusSetter <- R6Class(
         by <- private$by
       } else {
         check_string(by, allow_empty = FALSE, arg = ".by")
+        private$check_col(by)
       }
-      private$check_col(by)
       by
     }
   )
@@ -107,7 +107,7 @@ StatusSetterPlume <- R6Class(
   )
 )
 
-#' @title StatusSetterQuarto class
+#' @title StatusSetterPlumeQuarto class
 #' @description Internal class extending `StatusSetter` for `PlumeQuarto`.
 StatusSetterPlumeQuarto <- R6Class(
   classname = "StatusSetterPlumeQuarto",