+++ tests/testthat/test-get-contributions.R
@@ -33,6 +33,10 @@ test_that("get_contributions() return authors' contributions", {
     c("a: P.-P.P., R.R. and Z.Z.", "b: Z.Z.")
   )
   expect_equal(
+    aut$get_contributions(by_author = FALSE, sep = "; "),
+    c("a: Z.Z.; R.R. and P.-P.P.", "b: Z.Z.")
+  )
+  expect_equal(
     aut$get_contributions(by_author = FALSE, sep_last = " & "),
     c("a: Z.Z., R.R. & P.-P.P.", "b: Z.Z.")
   )
@@ -108,6 +112,19 @@ test_that("get_contributions() reorders CRediT roles alphabetically", {
   expect_equal(
     aut$get_contributions(),
     c("Formal analysis: R.R.", "Writing - original draft: Z.Z.")
+  )
+})
+
+test_that("author order is preserved when using CRediT and `by_author` (#50)", {
+  aut <- Plume$new(data.frame(
+    given_name = c("Z", "A"),
+    family_name = c("Z", "A"),
+    writing = c(1, NA),
+    analysis = c(NA, 1)
+  ))
+  expect_equal(
+    aut$get_contributions(roles_first = FALSE, by_author = TRUE),
+    c("Z.Z.: Writing - original draft", "A.A.: Formal analysis")
   )
 })